requirements.txt
geopandas
rasterio
dask
dask-jobqueue
distributed
jupyter
pytest
yapf
rasterstats
comet_ml
deepforest
